refactor: refactor whole project
