+++ src/routes/graphql/entities/Member/loader.ts
@@ -1,8 +1,9 @@
 import DataLoader from 'dataloader';
-import { Member } from '../types/model/ModelTypes.js';
-import { PrismaClient } from '@prisma/client';
 
-export const batchMemberTypeByUserIds =
+import { PrismaClient } from '@prisma/client';
+import { Member } from './types.js';
+
+const batchMemberTypeByUserIds =
   (prisma: PrismaClient) => async (userIds: readonly unknown[]) => {
     const ids = userIds as string[];
     const members: Member[] = await prisma.memberType.findMany({
