+++ src/routes/graphql/entities/Member/types.ts
@@ -1,0 +1,7 @@
+import { MemberTypeId } from '../../../member-types/schemas.js';
+
+export type Member = {
+  id: MemberTypeId | string;
+  discount: number;
+  postsLimitPerMonth: number;
+};
