+++ src/routes/graphql/entities/Post/post-type.ts
@@ -4,7 +4,7 @@ import {
   GraphQLObjectType,
   GraphQLString,
 } from 'graphql';
-import { UUIDType } from './uuid.js';
+import { UUIDType } from '../../types/uuid.js';
 
 export const PostType = new GraphQLObjectType({
   name: 'Post',
