+++ src/routes/graphql/index.ts
@@ -2,22 +2,7 @@ import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox';
 import { createGqlResponseSchema, gqlResponseSchema, schemaApp } from './schemas.js';
 import { graphql, parse, validate } from 'graphql';
 import depthLimit from 'graphql-depth-limit';
-import { createMemeberTypeLoader } from './loaders/membersLoader.js';
-import { createPostsLoader } from './loaders/postLoader.js';
-import { createProfileLoader } from './loaders/profileLoader.js';
-import { createSubscribedToUserLoader } from './loaders/subscribedToUser.js';
-import { createUserLoader } from './loaders/userLoader.js';
-import { createUserToSubscribeLoader } from './loaders/userSubscribedToLoader.js';
-import { PrismaClient } from '@prisma/client';
-
-const buldDataLoaders = (prisma: PrismaClient) => ({
-  postsLoader: createPostsLoader(prisma),
-  profileLoader: createProfileLoader(prisma),
-  userLoader: createUserLoader(prisma),
-  userSubscribedToLoader: createUserToSubscribeLoader(prisma),
-  subscribedToUser: createSubscribedToUserLoader(prisma),
-  memeberTypeLoader: createMemeberTypeLoader(prisma),
-});
+import { buldDataLoaders } from './loader.js';
 
 const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
   const { prisma } = fastify;
